requirements.txt
intel-tensorflow
opencv-python
Flask
easydict
tqdm
glog